Pass throwable to callback instead attach it to Route
Having throwable as part of Route made Route a mix between
being an immutable and having a state. It also opened up
an API to set an throwable on Route to change application
behaviour, that was never the intention for catching
throwable within Fram, it was to make it easier
for the callback react on exceptions.